# zip-repl-1/lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
  /** Optional machine-readable error code, for programmatic error handling. */
  code?: string;
}