# zip-repl-1/lib/api-zod/src/generated/types/stage1ResolvedParams.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * The fully resolved parameters actually used for the run (after defaults and auto-calibration).
 */
export interface Stage1ResolvedParams {
  cellSizeMeters: number;
  frequencyMHz: number;
  refractiveIndex: number;
  absorptionCoeff: number;
  sourceValue: number;
  minDb: number;
  pmlWidth: number;
  pmlMaxLoss: number;
  alpha3dBump: number;
  sourceXPercent: number;
  sourceYPercent: number;
  cellBudget: number;
}