# zip-repl-1/lib/api-zod/src/generated/types/stage1RunSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Stage1RunSummary {
  runId: string;
  createdAt: Date;
  gridRows: number;
  gridCols: number;
  peakDb: number;
  executionTimeMs: number;
}